SVN r5307 (sdl20)

+++ master/main.cpp
@@ -4,7 +4,7 @@
 // $Id$
 //
 // Copyright (C) 2000-2006 by Sergey Makovkin (CSDoom .62)
-// Copyright (C) 2006-2012 by The Odamex Team
+// Copyright (C) 2006-2015 by The Odamex Team.
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -36,6 +36,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+
+#include <stdint.h>
 
 #ifdef UNIX
 #include <netinet/in.h>
@@ -365,7 +367,7 @@ void pingServer(SServer &s)
 	}
 
 #ifdef _WIN32
-	s.key_sent = rand() * (int)GetModuleHandle(0) * time(0);
+	s.key_sent = rand() * (intptr_t)GetModuleHandle(0) * time(0);
 #else
 	s.key_sent = rand() * getpid() * time(0);
 #endif
